Refactor `TreeObserver`
The `TreeObserver` mechanism is currently only used by a handful of recipes and only via `TreeVisitor#collect()`. This use does however not require the diffing functionality as implemented using `ObjectDiffer`. This turns out to be rather expensive, so this commit makes it an opt-in feature and moves the logic from `TreeVisitor` into the `TreeObserver` itself.
