+++ rewrite-core/src/main/java/org/openrewrite/TreeVisitor.java
@@ -15,11 +15,6 @@
  */
 package org.openrewrite;
 
-import de.danielbechler.diff.ObjectDiffer;
-import de.danielbechler.diff.ObjectDifferBuilder;
-import de.danielbechler.diff.inclusion.Inclusion;
-import de.danielbechler.diff.inclusion.InclusionResolver;
-import de.danielbechler.diff.node.DiffNode;
 import io.micrometer.core.instrument.DistributionSummary;
 import io.micrometer.core.instrument.Metrics;
 import io.micrometer.core.instrument.Timer;
@@ -30,14 +25,12 @@ import org.openrewrite.internal.lang.Nullable;
 import org.openrewrite.marker.Marker;
 import org.openrewrite.marker.Markers;
 
-import java.beans.Transient;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.lang.reflect.TypeVariable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 
@@ -80,32 +73,6 @@ public abstract class TreeVisitor<T extends Tree, P> {
 
     private int visitCount;
     private final DistributionSummary visitCountSummary = DistributionSummary.builder("rewrite.visitor.visit.method.count").description("Visit methods called per source file visited.").tag("visitor.class", getClass().getName()).register(Metrics.globalRegistry);
-
-    private ObjectDiffer differ;
-
-    private ObjectDiffer getObjectDiffer() {
-        if (differ == null) {
-            differ = ObjectDifferBuilder.startBuilding()
-                    .inclusion()
-                    .resolveUsing(new InclusionResolver() {
-                        @Override
-                        public Inclusion getInclusion(DiffNode node) {
-                            if (node.getPropertyAnnotation(Transient.class) != null) {
-                                return Inclusion.EXCLUDED;
-                            }
-                            return Inclusion.DEFAULT;
-                        }
-
-                        @Override
-                        public boolean enablesStrictIncludeMode() {
-                            return false;
-                        }
-                    })
-                    .and()
-                    .build();
-        }
-        return differ;
-    }
 
     public boolean isAcceptable(SourceFile sourceFile, P p) {
         return true;
@@ -290,16 +257,7 @@ public abstract class TreeVisitor<T extends Tree, P> {
                     ExecutionContext ctx = (ExecutionContext) p;
                     for (TreeObserver.Subscription observer : ctx.getObservers()) {
                         if (observer.isSubscribed(tree)) {
-                            observer.getObserver().treeChanged(getCursor(), t);
-                            DiffNode diff = getObjectDiffer().compare(t, tree);
-                            AtomicReference<T> t2 = new AtomicReference<>(t);
-                            diff.visit((node, visit) -> {
-                                if (!node.hasChildren() && node.getPropertyName() != null) {
-                                    //noinspection unchecked
-                                    t2.set((T) observer.getObserver().propertyChanged(node.getPropertyName(), getCursor(), t2.get(), node.canonicalGet(tree), node.canonicalGet(t2.get())));
-                                }
-                            });
-                            t = t2.get();
+                            t = observer.treeChanged(getCursor(), t, tree);
                         }
                     }
                 }
